민팅 시 price 삭제, 판매 등록시 가격 입력 dialog 추가, NFT price Float으로 변경

+++ app/app/src/main/java/com/capstone/capstonenft/dto/LoginResponse.kt
@@ -19,7 +19,7 @@ data class Token(
     val imageSrc: String,
     val owner: String,
     val creator: String,
-    val price: Int?
+    val price: Float?
 ): Serializable
 
 data class TokenResponse(val token:Token)
